requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
pmdarima>=1.8.0
xgboost>=1.5.0
tensorflow>=2.8.0
keras>=2.8.0
yfinance>=0.1.70
fredapi>=0.5.0
jupyter>=1.0.0
pyyaml>=6.0.0
python-dotenv>=0.19.0
requests>=2.27.0
plotly>=5.5.0
tqdm>=4.62.0
python-dotenv>=1.0.1